refactor analysis package
